refac: calendar to modal
