+++ src/components/index.ts
@@ -8,6 +8,9 @@ import RoundInput from './RoundInput';
 import RoundSquareButton from './RoundSquareButton';
 import DefaultModal from './DefaultModal';
 import Calendar from './Calendar';
+import TextButton from './TextButton';
+import CalendarModal from './CalendarModal';
+import SelectCalendar from './SelectCalendar';
 
 export {
   Icons,
@@ -20,4 +23,7 @@ export {
   RoundSquareButton,
   DefaultModal,
   Calendar,
+  TextButton,
+  CalendarModal,
+  SelectCalendar,
 };
